+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -77,7 +77,6 @@ class TestTraitTest extends TestCase
              'assertIsBool' => true,
              'assertIsCallable' => fn() => '',
              'assertIsFloat' => 1.1,
-             'assertIsHtml' => '<b>html</b>',
              'assertIsInt' => 1,
              'assertIsIterable' => $Traversable,
              'assertIsJson' => '{"a":1,"b":2,"c":3,"d":4,"e":5}',
@@ -92,6 +91,21 @@ class TestTraitTest extends TestCase
             $staticCallable = [$this->TestCase, $assertMethod];
             forward_static_call($staticCallable, $value);
         }
+    }
+
+    /**
+     * `assertIsHtml()` is deprecated
+     * @group legacy
+     * @test
+     * @uses \Tools\TestSuite\TestTrait::__call()
+     * @uses \Tools\TestSuite\TestTrait::__callStatic()
+     */
+    public function testMagicCallAndCallStaticForAssertIsHtml(): void
+    {
+        $this->TestCase->assertIsHtml('<b>html</b>');
+        /** @var callable $staticCallable */
+        $staticCallable = [$this->TestCase, 'assertIsHtml'];
+        forward_static_call($staticCallable, '<b>html</b>');
     }
 
     /**
@@ -128,6 +142,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertArrayKeysEqual()
      */
@@ -149,6 +164,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertFileExtension()
      */
@@ -160,6 +176,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertFileMime()
      */
@@ -171,6 +188,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertImageSize()
      */
@@ -185,6 +203,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */
@@ -195,6 +214,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */
@@ -205,6 +225,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */
@@ -215,6 +236,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */
@@ -225,6 +247,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsMock()
      */
@@ -239,6 +262,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertObjectPropertiesEqual()
      */
@@ -255,6 +279,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertSameMethods()
      */
@@ -274,6 +299,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::createPartialMockForAbstractClass()
      */